+++ config/urls.py
@@ -14,10 +14,18 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
+from django.conf import settings
+from django.conf.urls.static import static
 from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
     path('admin/', admin.site.urls),
+    path('api/', include('expenses.api.urls')),
     path('', include('expenses.urls')),
 ]
+
+if settings.DEBUG:
+    # Serves uploaded profile pictures in development - a real
+    # deployment would have its web server (nginx etc.) do this instead.
+    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)